server/requirements.txt
boto3==1.35.12
Flask==3.1.2
flask_cors==6.0.1
python-dotenv==1.1.1
redis==5.0.8
gunicorn==23.0.0
